Fleshed out settings and metrics file
Now there is a separate file with different metrics for each pitch. The settings page now has a drop down for pitch type and can fill out the metrics information. I think we should make it a max and a min and scale how close the number is to the range. Also, need an option to select which hand to use with. Should LH and RH have its own file?