restore, new, get_avgs functionality
Added three new functions that can be called through HTTP requests. You can now restore the default base_metrics, update the base_metrics with new_metrics(), and get the averages of the metrics of a particular pitch. What is left is to add a category for each pitch called "avg" that has the averages for each metric. This will need to be done in default_metrics.json, then run the web_app and restore defaults to change it on the base_metrics.json file.